+++ blockchain-mediaart/src/components/Dashboard/Dashboard.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import 'bootstrap/dist/css/bootstrap.css';
+
+class Dashboard extends React.Component {
+    render(){
+        return (
+            <div>
+                <h1>Dashboard page</h1>
+                
+                <button type="button" class="btn btn-primary">Primary</button>
+                <button type="button" class="btn btn-secondary">Secondary</button>
+                <button type="button" class="btn btn-success">Success</button>
+                <button type="button" class="btn btn-danger">Danger</button>
+                <button type="button" class="btn btn-warning">Warning</button>
+                <button type="button" class="btn btn-info">Info</button>
+                <button type="button" class="btn btn-light">Light</button>
+                <button type="button" class="btn btn-dark">Dark</button>
+                <button type="button" class="btn btn-link">Link</button>
+            </div>
+        );
+    }
+}
+
+export default Dashboard;
