+++ blockchain-mediaart/src/components/Gallery/Gallery.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class Gallery extends React.Component {
+  render(){
+      return (
+          <div>Gallery page</div>
+      );
+  }
+}
+
+export default Gallery;
