+++ blockchain-mediaart/src/components/Main/Main.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class Main extends React.Component {
+  render(){
+      return (
+          <div>Main page</div>
+      );
+  }
+}
+
+export default Main;
